+++ demos/petstore/src/main/java/is/codion/demos/petstore/ui/TagItemEditPanel.java
@@ -43,6 +43,6 @@ public class TagItemEditPanel extends EntityEditPanel {
 	}
 
 	private TagEditPanel createTagEditPanel() {
-		return new TagEditPanel(new SwingEntityEditModel(Tag.TYPE, editModel().connection()));
+		return new TagEditPanel(new SwingEntityEditModel(Tag.TYPE, model().connection()));
 	}
 }